+++ R/data-mask.R
@@ -211,6 +211,12 @@ DataMask <- R6Class("DataMask",
 
       cols <- vec_c(cols_unused, cols_used)
 
+      # workaround until vctrs 0.3.5 is on CRAN
+      # (https://github.com/r-lib/vctrs/issues/1263)
+      if (length(cols) == 0) {
+        names(cols) <- character()
+      }
+
       # Match original ordering
       cols <- cols[across_vars]