+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Fixed `across()` issue where data frame columns would could not be referred to
   with `all_of()` in the nested case (`mutate()` within `mutate()`) (#5498).
+  
+* `across()` handles data frames with 0 columns (#5523). 
 
 # dplyr 1.0.2